+++ lib/mix/tasks/igniter.add.ex
@@ -37,13 +37,8 @@ defmodule Mix.Tasks.Igniter.Add do
     |> Enum.join(",")
     |> String.split(",")
     |> Enum.reduce(igniter, fn dep, igniter ->
-      case Igniter.Project.Deps.determine_dep_type_and_version(dep) do
-        {name, version} ->
-          Igniter.Project.Deps.add_dep(igniter, {name, version}, yes?: igniter.args.options[:yes])
-
-        :error ->
-          raise "Could not determine source for requested package #{dep}"
-      end
+      {name, version} = Igniter.Project.Deps.determine_dep_type_and_version!(dep)
+      Igniter.Project.Deps.add_dep(igniter, {name, version}, yes?: igniter.args.options[:yes])
     end)
     |> Igniter.add_task("deps.get")
   end